requirements.txt
streamlit
openai
python-dotenv
requests
beautifulsoup4
openai>=1.0.0